Binding Rows and Columns: In R, rbind() and cbind() functions are used to bind vectors or matrices row-wise and column-wise, respectively. They are often used to combine data structures horizontally or vertically.
Transposing Matrices:
The t() function in R transposes a matrix, interchanging its rows and columns. It is useful for rearranging data or preparing it for further processing.

Reshaping Data:
The matrix() function in R is used to reshape vectors or sequences into matrices with specified dimensions. It allows for organizing and structuring data in a desired format.